Temporarily remove harmonic minor scale
The mechanism for building chords currently uses OR logic to determine
whether the next valid interval within a scale is a major or a minor
third. In the harmonic minor scale, apparently there are notes where
both the major and minor thirds are in the scale. While I'd love to
keep the harmonic minor scale around (even though I never use it), it
probably is more pragmatic to remove it for now and spend some time
to sort this out afterwards.